+++ src/idtrackerai/video.py
@@ -132,7 +132,6 @@ class Video:
         roi_list: list[str] | None,
         use_bkg: bool,
         track_wo_identities: bool,
-        sigma_gaussian_blurring: float | None,
         check_segmentation: bool,
         identity_transfer: bool,
         knowledge_transfer_folder: Path | None,
@@ -146,8 +145,6 @@ class Video:
         video_path : str
             Path to a video file
         """
-        if sigma_gaussian_blurring is None:
-            sigma_gaussian_blurring = conf.SIGMA_GAUSSIAN_BLURRING
         if kwargs:
             logging.info(
                 f"Ignoring the next arguments in Video.__init__():\n{kwargs.keys()}"
@@ -165,7 +162,6 @@ class Video:
         self.number_of_animals = int(number_of_animals)
         """Number of animals in the video indicated by user"""
         self.set_video_paths(video_paths)
-        self.sigma_gaussian_blurring = sigma_gaussian_blurring
         self.data_policy: str = conf.DATA_POLICY
         self.frames_per_episode: int = conf.frames_per_episode
         self.version = metadata.version("idtrackerai")
@@ -621,6 +617,7 @@ class Video:
             break
         else:
             found = False
+            candidate_new_video_paths = []
             logging.error(f"Video file paths not found: {self.video_paths}")
 
         need_to_save = False
@@ -641,8 +638,10 @@ class Video:
             self.save()
 
     @staticmethod
-    def assert_video_paths(video_paths: Iterable[Path | str]):
-        accepted_extensions = conf.AVAILABLE_VIDEO_EXTENSION
+    def assert_video_paths(
+        video_paths: Iterable[Path | str], accepted_extensions: list[str] | None = None
+    ):
+        accepted_extensions = accepted_extensions or conf.AVAILABLE_VIDEO_EXTENSION
         assert video_paths, "Empty video_paths list"
 
         for path in video_paths: