+++ src/idtrackerai_start_app/segmentation_widgets/ROI_widget.py
@@ -1,6 +1,6 @@
 import numpy as np
 from cv2 import fitEllipse
-from qtpy.QtCore import Qt, Signal
+from qtpy.QtCore import Qt, Signal  # type: ignore
 from qtpy.QtGui import QColor, QPainterPath
 from qtpy.QtWidgets import (
     QCheckBox,