+++ tests/unit_test.py
@@ -1,10 +1,9 @@
 from importlib.resources import files
 
 import cv2
-import numpy as np
 import pytest
 
-from idtrackerai.animals_detection.segmentation import gaussian_blur, to_gray_scale
+from idtrackerai.animals_detection.segmentation import to_gray_scale
 
 TEST_VIDEO_SHAPE = (938, 1160)
 TEST_VIDEO_COMPRESSED_PATH_B = files("idtrackerai") / "data" / "test_B.avi"
@@ -30,17 +29,3 @@ def video_frame_0_gray(video_frame_0):
     assert gray.ndim == 2
     assert gray.shape == TEST_VIDEO_SHAPE
     return gray
-
-
-cases = [(None, "same"), (0, "same"), (10, "diff")]
-
-
-@pytest.mark.parametrize("sigma, expect", cases)
-def test_gaussian_blur(video_frame_0_gray, sigma, expect):
-    blurred_frame = gaussian_blur(video_frame_0_gray, sigma)
-    if expect == "same":
-        np.testing.assert_equal(video_frame_0_gray, blurred_frame)
-    else:  # expect == "diff"
-        np.testing.assert_raises(
-            AssertionError, np.testing.assert_equal, video_frame_0_gray, blurred_frame
-        )